Add typescript configurations in ./server. Edit Dockerfile in ./server (only for dev builds)
