+++ docker-compose.yaml
@@ -1,12 +1,5 @@
 version: '3'
 
-# The client service builds a Docker image from the ./client 
-# directory and maps port 3000 to the host machine. 
-# The server service builds a Docker image from the ./server 
-# directory and maps port 7036 to the host machine, and uses a volume
-# to persist data outside of the container.
-# The nginx service uses the official NGINX Docker image and maps port 80
-# to the host machine, and depends on the client and server services.
 services:
   client:
     build: ./client
