+++ server/Dockerfile
@@ -3,11 +3,10 @@ FROM node:14-alpine
 WORKDIR /app
 
 COPY package*.json ./
-
 RUN npm install
 
 COPY . .
+RUN npm run build
 
 EXPOSE 7036
-
-CMD [ "npm", "start" ]
+CMD ["npm", "start"]
