+++ server/package.json
@@ -2,10 +2,12 @@
     "name": "server",
     "version": "1.0.0",
     "description": "",
-    "main": "index.js",
+    "main": "build/server.js",
     "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1",
-        "start": "nodemon server.js"
+        "start": "node build/server.js",
+        "build": "tsc",
+        "dev": "nodemon build/server.ts",
+        "prettier": "prettier --write \"**/*.{js,jsx,ts,tsx,json,md}\""
     },
     "keywords": [],
     "author": "",
@@ -18,11 +20,11 @@
         "nodemon": "^2.0.22"
     },
     "devDependencies": {
-        "cors": "^2.8.5",
-        "dotenv": "^16.0.3",
-        "express": "^4.18.2",
-        "mongodb": "^5.1.0",
+        "@types/cors": "^2.8.13",
+        "@types/express": "^4.17.13",
+        "@types/node": "^18.15.11",
         "nodemon": "^2.0.22",
-        "prettier": "2.8.7"
+        "prettier": "2.8.7",
+        "typescript": "^4.5.5"
     }
 }
