--- server/server.js
@@ -1,18 +1,0 @@
-const express = require('express');
-const app = express();
-const cors = require('cors');
-require('dotenv').config({ path: './config.env' });
-const port = process.env.PORT || 7036;
-
-app.use(cors());
-app.use(express.json());
-app.use(require('./routes/record'));
-
-const dbo = require('./db/conn');
-
-app.listen(port, () => {
-    dbo.connectToServer(function (err) {
-        if (err) console.error(err);
-    });
-    console.log(`Server is running on port: ${port}`);
-});
